Added Sky Factory 2.2.3 Dockerfile.
Due to the nature of ATLauncher's backend, it seems that I can't use this
Dockerfile if the modpack updates its version. This is due to the use of a hash
that expires after a certain number of minutes, and is generated on the fly.

I will need to find a better way to build ATLauncher modpacks.