minecraft Dockerfile and script now use similar mechanics to spigot
This change is necessary to make sure that modpacks with NEI will
function properly.

It turns out that there is an existing bug in NEI wherein clients
will crash if level-name in server.properties is set to a path
instead of just 'world'.

Since that is no longer an option, I have no choice but to remove
the use of /var/lib/minecraft to separate the world data from
the game files, and use a similar mechanic to the Spigot Dockerfile
to sync files from the image into MINECRAFT_HOME.

I've added rsync, added a MINECRAFT_SRC variable for extending the
image, and set the level-name to world (cannot be overridden).

World templates are still possible with this image. The preferred
way is to copy world* folders to MINECRAFT_SRC.